[CORE] pkg "pronto_estimator_core" has been renamed "pronto_core"
