+++ pronto_core/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 2.8.3)
 
-project(pronto_estimator_core)
+project(pronto_core)
 add_compile_options(-std=c++11)
 set(PRONTO_CORE_LIB ${PROJECT_NAME})
