+++ pronto_core/README.md
