+++ pronto_core/include/pronto_core/definitions.hpp
