+++ pronto_core/include/pronto_core/pose_meas_module.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "pronto_estimator_core/sensing_module.hpp"
-#include "pronto_estimator_core/definitions.hpp"
+#include "pronto_core/sensing_module.hpp"
+#include "pronto_core/definitions.hpp"
 
 namespace MavStateEst {
