+++ pronto_core/include/pronto_core/sensing_module.hpp
@@ -1,6 +1,6 @@
 #pragma once
-#include "pronto_estimator_core/rbis_update_interface.hpp"
-#include "pronto_estimator_core/mav_state_est.hpp"
+#include "pronto_core/rbis_update_interface.hpp"
+#include "pronto_core/mav_state_est.hpp"
 
 namespace MavStateEst {
 template <class MessageT>
