+++ pronto_core/package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <package>
-  <name>pronto_estimator_core</name>
+  <name>pronto_core</name>
   <version>0.0.1</version>
   <description>
-    The pronto_estimator_core package contains the EKF and the main 
+    The pronto_core package contains the EKF and the main 
     modules to perform state estimation, including: inertial, gps, optical
     flow, etc.
   </description>
