+++ pronto_core/src/init_message_module.cpp
@@ -1,4 +1,4 @@
-#include "pronto_estimator_core/init_message_module.hpp"
+#include "pronto_core/init_message_module.hpp"
 
 namespace MavStateEst {
