+++ pronto_core/src/ins_module.cpp
@@ -1,4 +1,4 @@
-#include "pronto_estimator_core/ins_module.hpp"
+#include "pronto_core/ins_module.hpp"
 
 namespace MavStateEst {
