+++ pronto_core/src/pose_meas_module.cpp
@@ -1,4 +1,4 @@
-#include "pronto_estimator_core/pose_meas_module.hpp"
+#include "pronto_core/pose_meas_module.hpp"
 
 namespace MavStateEst {
