+++ pronto_core/src/scan_matcher_module.cpp
@@ -1,4 +1,4 @@
-#include "pronto_estimator_core/scan_matcher_module.hpp"
+#include "pronto_core/scan_matcher_module.hpp"
 #include "eigen_utils/eigen_rigidbody.hpp"
 
 namespace MavStateEst {
