+++ pronto_core/src/vicon_module.cpp
@@ -1,4 +1,4 @@
-#include "pronto_estimator_core/vicon_module.hpp"
+#include "pronto_core/vicon_module.hpp"
 
 namespace MavStateEst {
