+++ pronto_core/test/imu_module_test.cpp
@@ -1,4 +1,4 @@
-#include "pronto_estimator_core/ins_module.hpp"
+#include "pronto_core/ins_module.hpp"
 #include <random>
 #include <chrono>
